requirements.txt
requests==2.19.1
pycrypto==2.6.1
pandas==0.23.1
certifi==2018.4.16
chardet==3.0.4
idna==2.7
urllib3==1.23